+++ Sources/KeyboardKit/App/KeyboardApp.swift
@@ -25,7 +25,7 @@ import SwiftUI
 ///             bundleId: "com.keyboardkit.demo",
 ///             appGroupId: "group.com.keyboardkit.demo",
 ///             locales: [.english, .swedish, .persian],
-///             dictationDeepLink: "keyboardkit://dictation"
+///             deepLinks: .init(app: "keyboardkit://")
 ///         )
 ///     }
 /// }
@@ -49,7 +49,7 @@ public struct KeyboardApp {
     ///   - keyboardBundleId: The app's keyboard bundle identifier, by default `<bundleId>.keyboard`.
     ///   - appGroupId: The app's App Group identifier, if any.
     ///   - locales: The locales to use in the app, by default `.all`.
-    ///   - dictationDeepLink: The app's dictation deep link, if any.
+    ///   - deepLinks: App-specific deep links, if any.
     ///   - keyboardSettingsKeyPrefix: A custom keyboard settings key prefix, if any.
     public init(
         name: String,
@@ -58,7 +58,7 @@ public struct KeyboardApp {
         keyboardBundleId: String? = nil,
         appGroupId: String? = "",
         locales: [KeyboardLocale] = .all,
-        dictationDeepLink: String? = "",
+        deepLinks: DeepLinks? = nil,
         keyboardSettingsKeyPrefix: String? = nil
     ) {
         self.name = name
@@ -67,11 +67,12 @@ public struct KeyboardApp {
         self.keyboardBundleId = keyboardBundleId ?? "\(bundleId).keyboard"
         self.locales = locales
         self.licenseKey = licenseKey
+        self.deepLinks = deepLinks
         self.keyboardSettingsKeyPrefix = keyboardSettingsKeyPrefix
-        if let appGroupId, let dictationDeepLink {
+        if let appGroupId, let dictationLink = deepLinks?.dictation {
             dictationConfiguration = .init(
                 appGroupId: appGroupId,
-                appDeepLink: dictationDeepLink
+                appDeepLink: dictationLink
             )
         } else {
             dictationConfiguration = nil
@@ -96,20 +97,88 @@ public struct KeyboardApp {
     /// The locales to use in the app.
     public let locales: [KeyboardLocale]
 
-    /// The app's dictation deep link, if any.
+    /// App-specific deep links, if any.
+    public let deepLinks: DeepLinks?
+
+    /// The app's dictation configuration, if any.
     public let dictationConfiguration: Dictation.KeyboardConfiguration?
 
     /// A custom keyboard settings key prefix, if any.
     public let keyboardSettingsKeyPrefix: String?
+
+    @available(*, deprecated, message: "Use the deepLinks initializer instead.")
+    public init(
+        name: String,
+        licenseKey: String = "",
+        bundleId: String,
+        keyboardExtensionBundleId: String? = nil,
+        appGroupId: String? = nil,
+        locales: [KeyboardLocale] = .all,
+        dictationDeepLink: String
+    ) {
+        self.init(
+            name: name,
+            licenseKey: licenseKey,
+            bundleId: bundleId,
+            keyboardBundleId: keyboardExtensionBundleId,
+            appGroupId: appGroupId,
+            locales: locales,
+            deepLinks: .init(app: "", dictation: dictationDeepLink)
+        )
+    }
+
 }
 
 public extension KeyboardApp {
 
-    /// The keyboard extension bundle ID wildcard, which can
-    /// be used to see if any keyboard extension in this app
-    /// has been enabled.
-    var keyboardExtensionBundleIdWildcard: String {
+    /// This type can define app-specific deep links.
+    ///
+    /// You only have to provide an `app` url, and can leave
+    /// the rest blank to use standard URLs.
+    struct DeepLinks {
+
+        /// Create a custom keyboard deep links value.
+        ///
+        /// - Parameters:
+        ///   - app: A deep link for opening the app, e.g. `x://`.
+        ///   - dictation: A deep link for opening the app and starting dictation, by default `x://dictation`.
+        ///   - keyboardSettings: A deep link for opening the app's keyboard settings screen, by default `x://keyboardSettings`.
+        ///   - languageSettings: A deep link for opening the app's language settings screen, by default `x://languageSettings`.
+        ///   - themeSettings: A deep link for opening the app's theme settings screen, by default `x://themeSettings`.
+        public init(
+            app: String,
+            dictation: String? = nil,
+            keyboardSettings: String? = nil,
+            languageSettings: String? = nil,
+            themeSettings: String? = nil
+        ) {
+            self.app = app
+            self.dictation = dictation ?? "\(app)/dictation"
+            self.keyboardSettings = keyboardSettings ?? "\(app)/keyboardSettings"
+            self.languageSettings = languageSettings ?? "\(app)/languageSettings"
+            self.themeSettings = themeSettings ?? "\(app)/themeSettings"
+        }
+
+        public let app: String
+        public let dictation: String
+        public let keyboardSettings: String
+        public let languageSettings: String
+        public let themeSettings: String
+    }
+
+}
+
+public extension KeyboardApp {
+
+    /// This bundle ID wildcard that applies to all keyboard
+    /// extensions for the app.
+    var keyboardBundleIdWildcard: String {
         "\(bundleId).*"
+    }
+
+    @available(*, deprecated, renamed: "keyboardBundleIdWildcard")
+    var keyboardExtensionBundleIdWildcard: String {
+        keyboardBundleIdWildcard
     }
 }
 
@@ -121,8 +190,7 @@ private extension KeyboardApp {
             licenseKey: "abc123",
             bundleId: "com.keyboardkit.demo",
             appGroupId: "group.com.keyboardkit.demo",
-            locales: [.english, .swedish, .persian],
-            dictationDeepLink: "keyboardkit://dictation"
+            locales: [.english, .swedish, .persian]
         )
     }
 }