+++ RELEASE_NOTES.md
@@ -31,7 +31,7 @@ This version adds `KeyboardTheme` models from Pro to make it possible to add a n
 * `Image` has a new `keyboardTheme` image value.
 * `Keyboard.NumberPad` is a new number pad component.
 * `Keyboard.State` has a new `themeContext` property.
-* `KeyboardApp` has new `keyboardBundleId` properties.
+* `KeyboardApp` has new `deepLinks` and `keyboardBundleId` properties.
 * `KeyboardSettings` can now be set up for a `KeyboardApp`.
 * `KeyboardView` now applies next character probabilities to its items.
 * `KeyboardThemeContext` is a new context type for theme-related state.